refactor(cef): split state.rs auxiliary types into state/ submodules
Tier 1 item from docs/reports/REPORT_LARGE_FILE_MODULARIZATION_SCAN_2026_07_22.md.
~15 largely-independent type definitions (drag, window-meta, browser-pane,
pool, quit, top-level-creation-saga, UI-thread-gate) lived 1000+ lines above
the AppState struct that actually uses them as field types. Extracted into
state/{drag,window_meta,browser_pane,browser_handle,pool,quit,
top_level_creation,pending_reproject,ui_thread_gate}.rs, each re-exported
from state/mod.rs so every existing crate::state::X path is unchanged.
state/mod.rs keeps AppState + its Default/accessor impl + log_host_event.

Pure relocation, no logic changes. cargo build -p agentmux-cef clean;
cargo test -p agentmux-cef state:: — all 15 tests (pending_reproject +
ui_thread_gate, which already had embedded test modules) pass unchanged.

<!-- agentmux:agent_id=agentx -->
