+++ agentmux-cef/src/state/mod.rs
@@ -8,35 +8,31 @@ use parking_lot::Mutex;
 
 use cef::Browser;
 
-// ── Cross-window drag types (ported from src-tauri/src/state.rs) ─────────
+mod drag;
+mod window_meta;
+mod browser_pane;
+mod browser_handle;
+mod pool;
+mod quit;
+mod top_level_creation;
+mod pending_reproject;
+mod ui_thread_gate;
 
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-#[serde(rename_all = "lowercase")]
-pub enum DragType {
-    Pane,
-    Tab,
-}
-
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct DragPayload {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub block_id: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub tab_id: Option<String>,
-}
-
-#[derive(Debug, Clone, serde::Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct DragSession {
-    pub drag_id: String,
-    pub drag_type: DragType,
-    pub source_window: String,
-    pub source_workspace_id: String,
-    pub source_tab_id: String,
-    pub payload: DragPayload,
-    pub started_at: u64,
-}
+pub use drag::{DragType, DragPayload, DragSession};
+pub use window_meta::{WindowKind, WindowMeta, PendingWindowCreation};
+pub use browser_pane::{
+    BrowserPaneLifecycle, BrowserPaneEntry, PaneRect, WindowPlacement, PaneWindowState,
+};
+pub use browser_handle::{BrowserHandle, BrowserKind};
+pub use pool::{PoolState, PanePoolState};
+pub use quit::{QuitState, QuitReason};
+pub use top_level_creation::{
+    TopLevelCreationRequest, TopLevelSource, InFlightCreation, CreationPhase,
+    CompletedCreation, TopLevelCreationOutcome, TopLevelCreationState, EffectKind,
+    PendingBrowserPaneCreate, FloatingRedockGhostState,
+};
+pub use pending_reproject::PendingReprojectClosures;
+pub use ui_thread_gate::{UiThreadGate, MainReadyAction, SnapshotAction};
 
 // Phase B.5e — `WindowInstanceRegistry` struct deleted. Sequential
 // instance numbers are now owned by the launcher's reducer
@@ -57,460 +53,11 @@ pub struct BackendEndpoints {
     pub web_endpoint: String,
 }
 
-/// Window role in the AgentMux multi-window model.
-///
-/// Two distinct types with different taskbar treatment:
-/// - `FullInstance`: independent AgentMux window (like Chrome/VS Code new window).
-///   Appears in the Windows taskbar. All user-facing "new window" paths (status-bar
-///   version click, second `agentmux.exe` launch, `Ctrl+Shift+N`) create one.
-/// - `Subwindow`: hidden from the taskbar via `ITaskbarList::DeleteTab`. Only
-///   reachable through the backend `open_subwindow` API — reserved for agent /
-///   internal use cases (transient auxiliary views, tool-spawned panels). Closes
-///   when its parent full instance closes.
-#[derive(Copy, Clone, Debug, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum WindowKind {
-    FullInstance,
-    Subwindow,
-}
-
-/// Per-window metadata held alongside the CEF `Browser`. See `WindowKind` for
-/// the semantics of `kind` and `parent_instance_id`.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct WindowMeta {
-    pub label: String,
-    pub kind: WindowKind,
-    /// For `Subwindow` only: label of the `FullInstance` that owns this window.
-    /// `None` for `FullInstance`.
-    pub parent_instance_id: Option<String>,
-}
-
-/// Phase B.5 (window_meta step d) — pre-create handoff. Caller
-/// (`drag.rs::tear_off`, `commands/window.rs::open_new_window`,
-/// `window_pool.rs::spawn_pool_window`, `pane/creation.rs`) pushes
-/// one entry per window CEF is about to create; `client.rs::on_after_created`
-/// pops the head entry and uses `kind` for the Subwindow
-/// taskbar-hide branch + as the payload for `ReportWindowOpened`.
-///
-/// Replaces the previous `pending_window_labels: VecDeque<String>`
-/// queue + parallel caller-side `window_meta` writes that used to
-/// act as the kind/parent channel. Collapsing them into a single
-/// tuple eliminates the parallel-write race; on_after_created
-/// performs the single canonical `window_meta.insert` from the
-/// popped entry (kept as a synchronous host-side cache for
-/// open_subwindow's parent liveness check + cascade-close
-/// enumeration in `task dev` mode where launcher IPC is absent).
-#[derive(Clone, Debug)]
-pub struct PendingWindowCreation {
-    pub label: String,
-    pub kind: WindowKind,
-    pub parent_instance_id: Option<String>,
-}
-
-// ── Phase H — host reducer buildout ──────────────────────────────────────
-//
-// All types below are reducer-only state. PR #1 (h1-foundations) declares
-// them; subsequent PRs (#2-#5) wire callers through the reducer per the
-// a→b→c→d→e migration ratchet. See:
-//   docs/specs/SPEC_HOST_REDUCER_5PR_PLAN_2026-05-02.md
-//   docs/specs/SPEC_HOST_REDUCER_PHASE_H_2026-05-02.md
-//
-// These types intentionally have `#[allow(dead_code)]` because PR #1 ships
-// the scaffolding without callers — fields are populated by reducer arms but
-// no production code reads them yet. Subsequent PRs lift the allow as they
-// wire each migration.
-
-// ── Pane lifecycle (H.1) ─────────────────────────────────────────────────
-
-/// Lifecycle state of a browser pane (the `defwidget@browser` widget). Held
-/// inside `HostState.browser_panes` keyed by `block_id`. Mirrors the existing
-/// `PaneStateMachine::BrowserPaneLifecycle` (pane/lifecycle.rs:28); the existing
-/// type stays during PR #2's a→e migration. PR #2 step e deletes the
-/// pane/lifecycle.rs version and migrates all readers to this one.
-#[derive(Clone, Debug, PartialEq, Eq)]
-#[allow(dead_code)]
-pub enum BrowserPaneLifecycle {
-    /// Pane is alive and accepting operations (focus, resize, navigate).
-    Live,
-    /// Close requested; awaiting CEF on_before_close to fully tear down.
-    /// `since` carries the request timestamp for diagnostic purposes only;
-    /// nothing in the reducer is timer-driven.
-    Closing { since: std::time::Instant },
-}
-
-/// Per-pane reducer-managed entry. Replaces `pane::lifecycle::BrowserPaneEntry`
-/// (lifecycle.rs:42) at PR #2 step e.
-#[derive(Clone, Debug)]
-#[allow(dead_code)]
-pub struct BrowserPaneEntry {
-    pub block_id: String,
-    pub label: String,
-    pub lifecycle: BrowserPaneLifecycle,
-    /// The window this pane was created in (`main`, `floating-<uuid>`, …).
-    /// Used to detect a cross-window move (tear-off / redock): a create
-    /// request whose `window_label` differs from this must NOT be served by
-    /// re-navigating the existing browser in the OLD window (that leaves the
-    /// requested window black). See the `AlreadyLiveElsewhere` handling in
-    /// `reducer/panes.rs` + `browser_panes.rs`.
-    pub window_label: String,
-}
-
-// ── Pane window-placement state (pane-state reducer, Phase 0) ─────────────
-//
-// SPEC_PANE_STATE_REDUCER_2026-05-28.md (REVISION 2026-05-29 — folded into
-// HostState rather than a standalone PaneStateMachine, mirroring the Phase-H
-// consolidation that deleted `pane::lifecycle::PaneStateMachine`).
-//
-// This tracks the OS-window placement of a FLOATING pane (its
-// maximize/restore state + the rect to restore to). It is deliberately
-// SEPARATE from `BrowserPaneEntry`/`BrowserPaneLifecycle` above, which own
-// Live/Closing lifecycle: lifecycle stays in `HostState.browser_panes`;
-// placement lives in `HostState.pane_window_states`. Docked panes have NO
-// entry here — their "maximize" is backend magnify
-// (`LayoutState.magnifiednodeid`), routed by the frontend `<MaximizeButton>`
-// (spec §3.3a, b2), never through this reducer.
-
-/// Screen-space window rectangle in physical pixels. Distinct from
-/// `browser_api::Rect` (web/CSS coordinates) — this is for native window
-/// placement (the floating pane's outer HWND rect).
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Default)]
-#[allow(dead_code)]
-pub struct PaneRect {
-    pub left: i32,
-    pub top: i32,
-    pub right: i32,
-    pub bottom: i32,
-}
-
-/// OS-window placement for a floating pane. The shared maximize button's
-/// floating half toggles between `Normal` and `Maximized`; `Minimized` is
-/// OS-reported (Win+Down / system) and orthogonal — a minimized floater
-/// un-minimizes back to whichever of Normal/Maximized it held before, so
-/// the reducer remembers the pre-minimize placement.
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Default)]
-#[allow(dead_code)]
-pub enum WindowPlacement {
-    #[default]
-    Normal,
-    Maximized,
-    Minimized,
-}
-
-/// Per-floater window-placement entry, keyed by the floating-window LABEL
-/// (`floating-<uuid>`) in `HostState.pane_window_states`. Holds ONLY the
-/// floating window's OS placement and the rect to restore to after
-/// un-maximize. Keyed by label (not block_id) because floaters are tracked
-/// by window label everywhere (`window_hwnds`, the `?windowLabel=` URL, the
-/// `on_before_close` teardown) and are not in `browser_panes`.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-#[allow(dead_code)]
-pub struct PaneWindowState {
-    pub placement: WindowPlacement,
-    /// Rect to restore to after un-maximize / un-minimize. `None` until a
-    /// normal-mode rect has been observed (replaces the deleted
-    /// `AppState.floating_restored_rects` stash — spec §4).
-    pub last_known_normal_rect: Option<PaneRect>,
-}
-
-// ── Browser handle registry (H.2) ────────────────────────────────────────
-
-/// Wrapped CEF Browser handle stored in `HostState.browsers`. Replaces the
-/// raw `Mutex<HashMap<String, Browser>>` at `state.rs::AppState.browsers`
-/// at PR #2 step e.
-///
-/// `cef::Browser` is `Clone` (refcounted FFI handle) and safe to store
-/// inside the reducer's mutex-guarded state. Doesn't impl Debug, hence
-/// the manual `impl Debug` below for `BrowserHandle`.
-#[derive(Clone)]
-#[allow(dead_code)]
-pub struct BrowserHandle {
-    pub label: String,
-    pub browser: Browser,
-    pub kind: BrowserKind,
-    pub registered_at: std::time::Instant,
-}
-
-impl std::fmt::Debug for BrowserHandle {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("BrowserHandle")
-            .field("label", &self.label)
-            .field("kind", &self.kind)
-            .field("registered_at", &self.registered_at)
-            .field("browser", &"<cef::Browser>")
-            .finish()
-    }
-}
-
-/// Distinguishes top-level CEF Browsers (full-instance windows + pool
-/// windows) from pane CEF Browsers (children of a top-level). Determines
-/// taskbar treatment, lifecycle ownership, etc.
-#[derive(Clone, Debug)]
-#[allow(dead_code)]
-pub enum BrowserKind {
-    /// Top-level window (`main`, `window-*`, and the window pool). `is_pool=true`
-    /// while the window is in the warm window pool; cleared on promote. This is
-    /// the ONLY kind that keeps the instance alive — see `is_live_user_window`.
-    TopLevel { is_pool: bool },
-    /// Floating pane / tear-off in its own frameless `WS_POPUP` window —
-    /// `floating-<uuid>` (created directly) or `floating-pool-<uuid>` (taken from
-    /// the pane pool). `is_pool=true` while warm in the pane pool; cleared on
-    /// promote (`pane_pool.rs`). Floaters do **not** keep the instance alive
-    /// (invariant FP-LIFE) — `is_live_user_window` counts only
-    /// `TopLevel { is_pool: false }`, so a Floater is excluded by type, not by a
-    /// label-prefix string. They ARE trusted top-level renderers, so
-    /// `list_top_level_browsers` includes them for host JS-event emission.
-    Floater { is_pool: bool },
-    /// Browser pane child window. `block_id` correlates with the
-    /// `HostState.browser_panes` entry.
-    Pane { block_id: String },
-}
-
-// ── Pool state (H.4) ─────────────────────────────────────────────────────
-
-/// Pre-warmed window pool state. Replaces three separate fields on
-/// `AppState`: `window_pool: Mutex<VecDeque<String>>`,
-/// `unpromoted_pool_labels: Mutex<HashSet<String>>`, and
-/// `window_pool_respawn_in_flight: AtomicBool`. PR #3 migrates each
-/// caller through the a→e ratchet.
-#[derive(Default, Clone, Debug)]
-#[allow(dead_code)]
-pub struct PoolState {
-    /// Labels of pool windows whose renderer signaled ready (eligible
-    /// for promotion).
-    pub queue: std::collections::VecDeque<String>,
-    /// Labels spawned but not yet renderer-ready (and therefore not yet
-    /// in `queue`). Used for taskbar/exclusion filters during the spawn
-    /// → ready window.
-    pub unpromoted: std::collections::HashSet<String>,
-    /// Single-flight semaphore: true while a respawn task is in flight,
-    /// preventing stacked refills.
-    pub respawn_in_flight: bool,
-}
-
-/// Pre-warmed pane (floating) window pool state.
-/// Mirrors `PoolState` but for `floating-pool-{uuid}` frameless windows.
-#[derive(Default, Clone, Debug)]
-#[allow(dead_code)]
-pub struct PanePoolState {
-    pub queue: std::collections::VecDeque<String>,
-    pub unpromoted: std::collections::HashSet<String>,
-    pub respawn_in_flight: bool,
-}
-
-// ── Quit state (H.5) ─────────────────────────────────────────────────────
-
-/// Host process quit lifecycle. Replaces `is_quitting: AtomicBool` at
-/// `state.rs::AppState`. Three states; transitions are monotonic
-/// (Running → Draining → Quit, no regression).
-#[derive(Clone, Debug, PartialEq, Eq)]
-#[allow(dead_code)]
-pub enum QuitState {
-    /// Normal operation. All commands accepted (subject to per-arm rules).
-    Running,
-    /// `BeginDrain` dispatched. Pool refills suppressed; awaiting pool +
-    /// browsers to drain.
-    Draining { reason: QuitReason },
-    /// `ConfirmDrained` dispatched. Host quitting; no further commands.
-    Quit,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-#[allow(dead_code)]
-pub enum QuitReason {
-    /// User closed the last user-visible top-level window. Standard exit.
-    LastWindowClosed,
-    /// Launcher signaled HostShouldQuit (cross-process shutdown).
-    LauncherRequested,
-    /// External force-quit (Win32 WM_QUIT, signal, etc.).
-    External,
-}
-
-impl Default for QuitState {
-    fn default() -> Self { QuitState::Running }
-}
-
-// ── Top-level window creation runner (H.6) ───────────────────────────────
-
-/// A request to create a top-level window. Pushed to
-/// `HostState.top_level_creation.queue` via `EnqueueTopLevelWindow`.
-/// Carries the full spec the effect handler needs to call
-/// `ui_tasks::post_create_window`.
-#[derive(Clone, Debug)]
-#[allow(dead_code)]
-pub struct TopLevelCreationRequest {
-    pub label: String,
-    pub kind: WindowKind,
-    pub parent_instance_id: Option<String>,
-    pub url: String,
-    pub pos: (i32, i32),
-    pub size: (i32, i32),
-    pub frameless: bool,
-    /// `User`-initiated (fail-fast on contention) vs `Background` (pool
-    /// refill — may queue silently).
-    pub source: TopLevelSource,
-}
-
-/// Distinguishes user-facing creation requests (which fail-fast on
-/// contention) from background ones (pool refill — may queue indefinitely).
-#[derive(Clone, Debug, PartialEq, Eq)]
-#[allow(dead_code)]
-pub enum TopLevelSource {
-    /// Triggered by a user action (click "open new window", tear off a
-    /// tab, etc.). Reducer rejects with error if in-flight slot is
-    /// occupied — caller propagates a visible error to the frontend.
-    User,
-    /// Triggered by the runner itself (pool refill, recovery). Queues
-    /// behind in-flight; no caller waiting on completion.
-    Background,
-}
-
-/// The single in-flight top-level creation. Singleton invariant enforced
-/// by the reducer (at most one Some across all action sequences).
-///
-/// **No `deadline` field. No watchdog.** The reducer reacts only to
-/// observable CEF callbacks: `on_after_created` (success),
-/// `on_render_process_terminated` (renderer crash), `on_before_close`
-/// (cancel mid-create). If none fire, the slot stays occupied
-/// permanently — user-initiated creates fail-fast with visible error
-/// per the no-timer directive.
-#[derive(Clone, Debug)]
-#[allow(dead_code)]
-pub struct InFlightCreation {
-    pub creation_id: u64,
-    pub label: String,
-    pub started_at: std::time::Instant,
-    pub phase: CreationPhase,
-}
-
-/// Phase progression of an in-flight top-level creation. Monotonic;
-/// `AdvanceCreationPhase` (if added later) refuses regression.
-#[derive(Clone, Debug, Copy, PartialEq, Eq, PartialOrd, Ord)]
-#[repr(u8)]
-#[allow(dead_code)]
-pub enum CreationPhase {
-    /// Effect handler has dispatched `post_create_window`; CEF has not
-    /// yet fired any callback.
-    Started = 0,
-    /// CEF `on_after_created` fired — Browser exists, renderer alive.
-    BrowserCallbackFired = 1,
-}
-
-/// Archived completion record for the runner's history ring buffer.
-/// Bounded at `TOP_LEVEL_CREATION_HISTORY_CAP` (50); oldest evicted FIFO.
-#[derive(Clone, Debug)]
-#[allow(dead_code)]
-pub struct CompletedCreation {
-    pub creation_id: u64,
-    pub label: String,
-    pub outcome: TopLevelCreationOutcome,
-    pub started_at: std::time::Instant,
-    pub finished_at: std::time::Instant,
-    pub last_phase: CreationPhase,
-}
-
-/// Why a top-level creation completed. `Completed` is happy-path; the
-/// other variants are observable failure modes.
-#[derive(Clone, Debug)]
-#[allow(dead_code)]
-pub enum TopLevelCreationOutcome {
-    /// CEF `on_after_created` fired; browser registered. Normal completion.
-    Completed,
-    /// CEF `on_render_process_terminated` fired during creation. Renderer
-    /// process died (crash, OOM, killed).
-    RendererTerminated { status: String },
-    /// CEF `on_before_close` fired during creation. Browser closed
-    /// externally (user action, parent close, etc.).
-    ExternallyClosed,
-}
-
-/// Reducer-managed runner state. Owns the queue, in-flight slot, history,
-/// and id allocator.
-#[derive(Default, Clone, Debug)]
-#[allow(dead_code)]
-pub struct TopLevelCreationState {
-    pub queue: std::collections::VecDeque<TopLevelCreationRequest>,
-    pub in_flight: Option<InFlightCreation>,
-    pub history: std::collections::VecDeque<CompletedCreation>,
-    pub next_creation_id: u64,
-}
-
-// ── Effects (carrier for side-effect-bearing events) ─────────────────────
-
-/// Side-effect descriptor emitted by reducer arms. Carried inside
-/// `HostEvent::Effect(EffectKind)`. The effects executor in
-/// `AppState::host_dispatch_with_effects` dispatches each kind to the
-/// appropriate imperative handler (e.g., posting a CEF UI task).
-///
-/// Reducer arms emit effects but never execute them; this preserves the
-/// pure-functional discipline of `update()`. Manual Debug impl below
-/// because `cef::Browser` doesn't impl Debug.
-#[derive(Clone)]
-#[allow(dead_code)]
-pub enum EffectKind {
-    /// Begin top-level window creation by posting `ui_tasks::post_create_window`.
-    /// Carried by `HostEvent::TopLevelCreationStarted`'s effect path.
-    PostCreateWindow { request: TopLevelCreationRequest, creation_id: u64 },
-    /// Spawn a pool window (PR #3 wires this when pool drops below TARGET_SIZE).
-    SpawnPoolWindow,
-    /// Close an orphan CEF browser whose label doesn't match any in-flight
-    /// or registered entry. Used by H.6's mismatched-callback handler to
-    /// prevent label collision.
-    CloseOrphanBrowser { browser: Browser },
-}
-
-impl std::fmt::Debug for EffectKind {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            EffectKind::PostCreateWindow { request, creation_id } => f
-                .debug_struct("PostCreateWindow")
-                .field("creation_id", creation_id)
-                .field("label", &request.label)
-                .finish(),
-            EffectKind::SpawnPoolWindow => f.write_str("SpawnPoolWindow"),
-            EffectKind::CloseOrphanBrowser { .. } => f
-                .debug_struct("CloseOrphanBrowser")
-                .field("browser", &"<cef::Browser>")
-                .finish(),
-        }
-    }
-}
-
 /// Shared application state for the CEF host.
 ///
 /// Unlike the Tauri version, this uses `Arc<AppState>` directly instead of
 /// `tauri::State<AppState>`. The sidecar child is `std::process::Child` instead
 /// of `tauri_plugin_shell::process::CommandChild`.
-
-/// A browser-pane create deferred while the block_id was still `Closing`
-/// (old CEF Browser mid-teardown). Owned by the reducer's
-/// `HostState.pending_browser_pane_creates` (NOT `AppState`) so stash-on-`Closing`
-/// and remove-on-close are atomic under the single host_state lock. The
-/// close-completion arms (`CompleteBrowserPaneClose`/`DrainBrowserPaneByLabel`)
-/// hand it back via `DispatchOutput.pending_browser_pane_create_to_replay`;
-/// the IPC handler replays it (now `Fresh`). The deterministic
-/// re-create-after-close that fixes the redock "pane sometimes won't load"
-/// race (no frontend retry / timer). Rect stored as raw i32s to keep this
-/// type independent of `cef::Rect`.
-#[derive(Clone, Debug)]
-pub struct PendingBrowserPaneCreate {
-    pub url: String,
-    pub x: i32,
-    pub y: i32,
-    pub width: i32,
-    pub height: i32,
-    pub window_label: String,
-}
-
-/// Phase 4b — ghost state stored per target-window label during a floating-pane
-/// redock drag. The target renderer pushes `block_id + dir` when it shows the
-/// ghost overlay; the floater reads it at drop time to pass a directional hint
-/// to the `RedockFloatingPane` saga.
-#[derive(Clone, Debug)]
-pub struct FloatingRedockGhostState {
-    pub block_id: String,
-    pub dir: u8,
-}
-
 pub struct AppState {
     // Phase B.5 (window_id_map step e) — `window_id_map` field
     // deleted. Authoritative copy lives in the launcher's
@@ -979,310 +526,6 @@ pub struct AppState {
     /// (safe, if imperfect) fallback behavior as before this whole cleanup
     /// existed: the old data lingers, but is never lost.
     pub pending_reproject_closures: Mutex<PendingReprojectClosures>,
-}
-
-/// See `AppState::pending_reproject_closures`. A thin, named wrapper around
-/// `HashMap<new_label, old_window_id>` rather than a bare map — reagent P2
-/// (PR #2032, 2026-07-08) noted the deferred-close logic had no automated
-/// test coverage; giving it named methods makes the intent explicit and
-/// gives something to unit-test directly, without needing a live instance.
-#[derive(Debug, Default)]
-pub struct PendingReprojectClosures(std::collections::HashMap<String, String>);
-
-impl PendingReprojectClosures {
-    /// Record that `new_label`'s eventual `register_backend_window` call
-    /// should trigger closing `old_window_id`. Called right after
-    /// `reproject_from_snapshot` returns — before any confirmation the new
-    /// window actually exists.
-    pub fn stage(&mut self, new_label: String, old_window_id: String) {
-        self.0.insert(new_label, old_window_id);
-    }
-
-    /// Called from `register_backend_window` for every label that
-    /// registers. Returns the old window_id to close if `label` was staged
-    /// (removing the entry so it fires at most once); `None` for any
-    /// ordinary (non-reprojected) window registration.
-    pub fn confirm(&mut self, label: &str) -> Option<String> {
-        self.0.remove(label)
-    }
-
-    #[cfg(test)]
-    fn len(&self) -> usize {
-        self.0.len()
-    }
-}
-
-#[cfg(test)]
-mod pending_reproject_closures_tests {
-    use super::*;
-
-    #[test]
-    fn confirm_returns_none_for_unstaged_label() {
-        let mut p = PendingReprojectClosures::default();
-        assert_eq!(p.confirm("window-abc"), None);
-    }
-
-    #[test]
-    fn confirm_returns_the_staged_old_id() {
-        let mut p = PendingReprojectClosures::default();
-        p.stage("window-new".to_string(), "old-srv-id".to_string());
-        assert_eq!(p.confirm("window-new"), Some("old-srv-id".to_string()));
-    }
-
-    #[test]
-    fn confirm_only_fires_once() {
-        let mut p = PendingReprojectClosures::default();
-        p.stage("window-new".to_string(), "old-srv-id".to_string());
-        assert_eq!(p.confirm("window-new"), Some("old-srv-id".to_string()));
-        // A second registration for the same label (shouldn't normally
-        // happen, but must not re-trigger a close) gets nothing.
-        assert_eq!(p.confirm("window-new"), None);
-    }
-
-    #[test]
-    fn confirm_is_independent_per_label() {
-        let mut p = PendingReprojectClosures::default();
-        p.stage("window-a".to_string(), "old-a".to_string());
-        p.stage("window-b".to_string(), "old-b".to_string());
-        assert_eq!(p.len(), 2);
-        assert_eq!(p.confirm("window-a"), Some("old-a".to_string()));
-        assert_eq!(p.len(), 1);
-        assert_eq!(p.confirm("window-b"), Some("old-b".to_string()));
-        assert_eq!(p.len(), 0);
-    }
-
-    #[test]
-    fn unrelated_label_registration_is_a_noop() {
-        // e.g. a pool window, or main itself — never staged, must not panic
-        // or spuriously return something.
-        let mut p = PendingReprojectClosures::default();
-        p.stage("window-real-reproject".to_string(), "old-id".to_string());
-        assert_eq!(p.confirm("window-pool-abc123"), None);
-        assert_eq!(p.confirm("main"), None);
-        // The real one is still there, untouched.
-        assert_eq!(p.confirm("window-real-reproject"), Some("old-id".to_string()));
-    }
-}
-
-/// See `AppState::ui_thread_gate`.
-#[derive(Default)]
-pub struct UiThreadGate {
-    pub ready: bool,
-    pub stashed: Option<Vec<agentmux_common::ipc::WindowSnapshot>>,
-    /// SPEC_PILLAR1_STEP4 Phase 3 addendum (reagent P1, PR #2032, 2026-07-08)
-    /// — the same `Event::Snapshot`'s sibling `backend_window_ids` field
-    /// (label → real srv window_id), stashed alongside `stashed` under the
-    /// same lock acquisition so the two never drift apart. Needed so the
-    /// fast-path replay can stage a deferred close for each recreated
-    /// window via `reproject_from_snapshot_and_stage_closures` — without
-    /// this, the fast path had no way to learn the OLD window's real srv
-    /// id (`WindowSnapshot.label` alone is only the launcher's in-memory
-    /// label), so `Client.windowids` grew unboundedly on every ordinary
-    /// (launcher-survives) crash.
-    pub stashed_backend_window_ids: Option<Vec<(String, String)>>,
-    /// SPEC_PILLAR1_STEP4 Phase 3 — set true the moment EITHER reproject
-    /// path (fast, from the launcher's snapshot; or slow, from srv) has
-    /// actually been triggered, so only one of them ever creates windows.
-    /// Without this, a fast-path snapshot arriving late (after the slow
-    /// path already ran, or vice versa) would run a second, redundant
-    /// reproject on top of the first.
-    pub reprojected: bool,
-    /// SPEC_PILLAR1_STEP4 Phase 3 addendum (reagent P0, PR #2017,
-    /// 2026-07-08) — set true by `"main"`'s registration when no fast-path
-    /// stash was available, meaning the slow path SHOULD run but can't yet:
-    /// `reproject_from_srv` needs `"main"`'s own confirmed srv `window_id`
-    /// to know which entry in `Client.windowids` to exclude/treat as the
-    /// parent-linkage sentinel, and that isn't known until `"main"`'s own
-    /// `register_backend_window` call lands (well after native browser
-    /// creation — the frontend has to load and bootstrap first). The
-    /// original design used `windowids[0]` as a stand-in, positionally —
-    /// reagent caught that `windowids` gets reordered by `focus_window`
-    /// (`agentmux-srv/.../wcore/window.rs:164`) to put the last-focused
-    /// window at index 0, which is not reliably `"main"`. Consumed (cleared)
-    /// by whichever of {`register_backend_window`'s `"main"` branch, a
-    /// late-arriving real fast-path snapshot} runs first — see both call
-    /// sites' comments for why exactly one of them wins.
-    pub pending_slow_path: bool,
-}
-
-/// What the caller of `UiThreadGate::on_main_ready` should do.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum MainReadyAction {
-    /// A real fast-path snapshot was stashed — replay it now.
-    ReplayFastPath,
-    /// Nothing useful was stashed — the caller should await the slow path
-    /// (triggered later, elsewhere, once `"main"`'s srv id is known).
-    AwaitSlowPath,
-    /// Something already reprojected (shouldn't normally happen — `"main"`
-    /// only registers once — kept as a defensive no-op, not a panic).
-    Noop,
-}
-
-/// What the caller of `UiThreadGate::on_snapshot` should do.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum SnapshotAction {
-    /// UI thread isn't ready yet — stash the snapshot for later replay.
-    Stash,
-    /// Run the fast path now with this snapshot's data.
-    RunFastPath,
-    /// Something already reprojected — ignore this snapshot.
-    Skip,
-}
-
-impl UiThreadGate {
-    /// SPEC_PILLAR1_STEP4 Phase 3 — reagent P2 (PR #2017, 2026-07-08): this
-    /// state machine has had three real bugs found across three review
-    /// rounds (a TOCTOU, a "stash exists" vs "stash has data" conflation,
-    /// and a positional-vs-value identity bug), each only caught by manual
-    /// live-kill verification. Extracted into pure, directly unit-testable
-    /// methods so future edits have more than that to lean on. Callers
-    /// (`client/lifecycle.rs`, `launcher_ipc.rs`, `commands/window/meta.rs`)
-    /// still own all I/O (taking the stash's data, actually calling
-    /// `reproject_from_snapshot`/`reproject_from_srv`) — these methods only
-    /// decide, they never act.
-    ///
-    /// Called once, when `"main"` registers. `has_extra` = does the
-    /// (already-taken) stash contain anything beyond `"main"` itself? A
-    /// stash existing is NOT the same as it having anything useful — see
-    /// the call site for why that distinction mattered.
-    pub fn on_main_ready(&mut self, has_extra: bool) -> MainReadyAction {
-        self.ready = true;
-        if has_extra {
-            self.reprojected = true;
-            MainReadyAction::ReplayFastPath
-        } else if !self.reprojected {
-            self.pending_slow_path = true;
-            MainReadyAction::AwaitSlowPath
-        } else {
-            MainReadyAction::Noop
-        }
-    }
-
-    /// Called whenever an `Event::Snapshot` arrives from the launcher.
-    pub fn on_snapshot(&mut self) -> SnapshotAction {
-        if !self.ready {
-            SnapshotAction::Stash
-        } else if self.reprojected {
-            SnapshotAction::Skip
-        } else {
-            // ready && !reprojected: a slow path is pending but hasn't
-            // fired yet. Real fast-path data is strictly better (has
-            // `last_rect`, no srv round-trip) — it wins the race.
-            self.reprojected = true;
-            self.pending_slow_path = false;
-            SnapshotAction::RunFastPath
-        }
-    }
-
-    /// Called from `register_backend_window`'s `"main"` branch, once
-    /// `"main"`'s own confirmed srv `window_id` is known. Returns whether
-    /// the caller should now run the slow path.
-    pub fn on_main_backend_window_registered(&mut self) -> bool {
-        if self.pending_slow_path && !self.reprojected {
-            self.pending_slow_path = false;
-            self.reprojected = true;
-            true
-        } else {
-            false
-        }
-    }
-}
-
-#[cfg(test)]
-mod ui_thread_gate_tests {
-    use super::*;
-
-    #[test]
-    fn main_ready_with_extra_replays_fast_path_and_sets_reprojected() {
-        let mut gate = UiThreadGate::default();
-        let action = gate.on_main_ready(true);
-        assert_eq!(action, MainReadyAction::ReplayFastPath);
-        assert!(gate.ready);
-        assert!(gate.reprojected);
-        assert!(!gate.pending_slow_path);
-    }
-
-    #[test]
-    fn main_ready_without_extra_awaits_slow_path() {
-        let mut gate = UiThreadGate::default();
-        let action = gate.on_main_ready(false);
-        assert_eq!(action, MainReadyAction::AwaitSlowPath);
-        assert!(gate.ready);
-        assert!(!gate.reprojected);
-        assert!(gate.pending_slow_path);
-    }
-
-    #[test]
-    fn main_ready_is_noop_if_already_reprojected() {
-        // Defensive case — shouldn't happen in practice ("main" registers
-        // once) but must not double-trigger if it somehow did.
-        let mut gate = UiThreadGate::default();
-        gate.reprojected = true;
-        let action = gate.on_main_ready(false);
-        assert_eq!(action, MainReadyAction::Noop);
-        assert!(!gate.pending_slow_path);
-    }
-
-    #[test]
-    fn snapshot_before_ready_stashes() {
-        let mut gate = UiThreadGate::default();
-        assert_eq!(gate.on_snapshot(), SnapshotAction::Stash);
-        assert!(!gate.reprojected);
-    }
-
-    #[test]
-    fn snapshot_after_ready_with_pending_slow_path_wins_and_cancels_it() {
-        let mut gate = UiThreadGate::default();
-        gate.on_main_ready(false); // ready=true, pending_slow_path=true
-        assert!(gate.pending_slow_path);
-        let action = gate.on_snapshot();
-        assert_eq!(action, SnapshotAction::RunFastPath);
-        assert!(gate.reprojected);
-        assert!(!gate.pending_slow_path);
-    }
-
-    #[test]
-    fn snapshot_after_reprojected_is_skipped() {
-        let mut gate = UiThreadGate::default();
-        gate.on_main_ready(true); // reprojected=true immediately
-        let action = gate.on_snapshot();
-        assert_eq!(action, SnapshotAction::Skip);
-    }
-
-    #[test]
-    fn slow_path_runs_when_pending_and_not_yet_reprojected() {
-        let mut gate = UiThreadGate::default();
-        gate.on_main_ready(false); // pending_slow_path=true
-        assert!(gate.on_main_backend_window_registered());
-        assert!(gate.reprojected);
-        assert!(!gate.pending_slow_path);
-    }
-
-    #[test]
-    fn slow_path_does_not_run_twice() {
-        let mut gate = UiThreadGate::default();
-        gate.on_main_ready(false);
-        assert!(gate.on_main_backend_window_registered());
-        // Second call (e.g. a duplicate register_backend_window) must not
-        // re-trigger the slow path.
-        assert!(!gate.on_main_backend_window_registered());
-    }
-
-    #[test]
-    fn slow_path_does_not_run_if_fast_path_already_won() {
-        let mut gate = UiThreadGate::default();
-        gate.on_main_ready(false); // pending_slow_path=true
-        gate.on_snapshot(); // fast path wins, cancels pending_slow_path
-        assert!(!gate.on_main_backend_window_registered());
-    }
-
-    #[test]
-    fn slow_path_does_not_run_if_never_pending() {
-        // Fast path had data from the start — pending_slow_path was never set.
-        let mut gate = UiThreadGate::default();
-        gate.on_main_ready(true);
-        assert!(!gate.on_main_backend_window_registered());
-    }
 }
 
 impl Default for AppState {
